requirements.txt
flask-sqlalchemy==2.3.2
elasticsearch==5.2.0
numpy==1.14.3
scipy==1.0.0
pandas==0.21.1
matplotlib==2.2.2
scikit-learn==0.19.1
# pip install PyJWT==1.5.3
PyJWT==1.5.3
#pandasticsearch
seaborn==0.8
#kafka-python
#Flask-Script
#Flask-Migrate
#flask


# =================数据库
# cx_oracle 使用5.1.2
cx_oracle==6.4.1
pymssql==2.1.3
# pip install MySQL-python
# conda install mysql-python
# MySQLdb 只适用于python2.x，发现pip装不上。它在py3的替代品是：import pymysql
# python2 MySQL-python
# python3 pymysql
# pip install PyMySQL



#决策树可视化环境搭建
#scikit-learn中决策树的可视化一般需要安装graphviz。主要包括graphviz的安装和python的graphviz插件的安装。
#第一步是安装graphviz。下载地址在：http://www.graphviz.org/。如果你是linux，可以用apt-get或者yum的方法安装。如果是windows，就在官网下载msi文件安装。无论是linux还是windows，装完后都要设置环境变量，将graphviz的bin目录加到PATH，比如我是windows，将C:/Program Files (x86)/Graphviz2.38/bin/加入了PATH
#第二步是安装python插件graphviz： pip install graphviz
#第三步是安装python插件pydotplus。这个没有什么好说的: pip install pydotplus
#这样环境就搭好了，有时候python会很笨，仍然找不到graphviz，这时，可以在代码里面加入这一行：
#os.environ["PATH"] += os.pathsep + 'C:/Program Files (x86)/Graphviz2.38/bin/'
#注意后面的路径是你自己的graphviz的bin目录。


# ============================================================
#python 3.6
# facets
protobuf==3.2.0


nltk==3.3